refactor: move utils.py to data/normalization.py
- Rename `src/mlcast/utils.py` to `src/mlcast/data/normalization.py`
- Update all references in `zarr_dataset.py`, `convgru.py`, and tests
