+++ logs/mlcast/version_0/config.yaml
@@ -1,0 +1,58 @@
+__fn__: mlcast.configs.Experiment
+pl_module:
+  __fn__: mlcast.models.convgru.RadarLightningModel
+  input_channels: 1
+  num_blocks: 5
+  ensemble_size: 2
+  noisy_decoder: false
+  forecast_steps: 12
+  loss_class: crps
+  loss_params:
+    temporal_lambda: 0.01
+  masked_loss: true
+  optimizer_class: torch.optim.adam.Adam
+  optimizer_params:
+    lr: 0.0001
+    fused: true
+  lr_scheduler_class: torch.optim.lr_scheduler.ReduceLROnPlateau
+  lr_scheduler_params:
+    mode: min
+    factor: 0.5
+    patience: 10
+data:
+  __fn__: mlcast.data.zarr_datamodule.RadarDataModule
+  zarr_path: ./data/radar.zarr
+  csv_path: ./data/sampled_datacubes.csv
+  variable_name: RR
+  steps: 18
+  train_ratio: 0.7
+  val_ratio: 0.15
+  return_mask: true
+  deterministic: false
+  augment: true
+  batch_size: 16
+  num_workers: 8
+  pin_memory: true
+trainer:
+  __fn__: pytorch_lightning.trainer.trainer.Trainer
+  accelerator: auto
+  logger:
+    __fn__: pytorch_lightning.loggers.tensorboard.TensorBoardLogger
+    save_dir: logs
+    name: mlcast
+  callbacks:
+  - __fn__: pytorch_lightning.callbacks.model_checkpoint.ModelCheckpoint
+    monitor: val_loss
+    save_top_k: 1
+    mode: min
+  - __fn__: pytorch_lightning.callbacks.model_checkpoint.ModelCheckpoint
+    monitor: train_loss_epoch
+    save_top_k: 1
+    mode: min
+  - __fn__: pytorch_lightning.callbacks.early_stopping.EarlyStopping
+    monitor: val_loss
+    patience: 100
+    mode: min
+  - __fn__: pytorch_lightning.callbacks.lr_monitor.LearningRateMonitor
+    logging_interval: step
+  max_epochs: 10
