+++ plan.md
@@ -5,10 +5,10 @@
 - [x] **Test Fixtures**: Create `tests/conftest.py` with a session-scoped fixture `italian_dataset` that explicitly downloads and caches the Italian DPC dataset from S3 locally, restricted to the first 100 timesteps using `.isel(time=slice(0, 100))` to ensure offline compatibility and avoid `simplecache` corruption bugs.
 
 ## Phase 1: Package Restructuring & Cleanup (PR Feedback)
-- [ ] **Tests**: Create `tests/data/test_normalization.py` and `tests/test_losses.py`.
-   - [ ] Verify `rainrate_to_normalized` and its inverse.
-   - [ ] Verify `AFCRPS` and `CRPS` loss shapes.
-- [ ] **Rename & Move**: Move `src/mlcast/utils.py` to `src/mlcast/data/normalization.py`. Update all imports.
+- [x] **Tests**: Create `tests/data/test_normalization.py` and `tests/test_losses.py`.
+   - [x] Verify `rainrate_to_normalized` and its inverse.
+   - [x] Verify `AFCRPS` and `CRPS` loss shapes.
+- [x] **Rename & Move**: Move `src/mlcast/utils.py` to `src/mlcast/data/normalization.py`. Update all imports.
 - [ ] **Normalization Registry**: Create a `NORMALIZATION_REGISTRY` dict in `normalization.py` mapping CF standard names (e.g., `"rainfall_rate"`) to their respective normalization functions.
 - [ ] **Visualization**: Extract `apply_radar_colormap` and `log_images` out of the model file into a new `src/mlcast/visualization.py`.
 - [ ] **Losses (`src/mlcast/losses.py`)**:
